[#25] DateSeparatorGenerator 구현
